MYLUTECE-58 : Fix MyLuteceAppTest tests
timestamp in MySQL has a resolution of a second. Connection logs are searched with strict inequalities
on the timestamps. As the tests run quite fast, the connection log search happens in the same second as
the entry was made, hence the failure.

To fix the test, wait a second after setting up the connection log.
Refactor common code into setUp and tearDown.
Check the precondition that the connection log inserted can be found.

Another solution might have been to search the connection log with <= and >=, but that migth break some
assumptions